+++ Runtime/Scripts/Utilities/AssetUtility.cs
@@ -11,7 +11,8 @@ internal static class AssetUtility {
     /// <summary>
     /// Normalize an absolute path under Unity project to make it relative to the Unity project folder.
     /// Paths that are outside Unity project will be unchanged.
-    /// Only slash ('/') is regarded as a directory separator. 
+    /// Will always return with directory separators using slash ('/'), but can handle both slash/backslash
+    /// as the directory separator for input. 
     /// Ex: C:/TempUnityProject/Assets/Foo.prefab => Assets/Foo.prefab
     ///     C:/NonUnityProject/Foo.prefab => C:/NonUnityProject/Foo.prefab
     /// </summary>
@@ -21,11 +22,44 @@ internal static class AssetUtility {
         if (string.IsNullOrEmpty(path))
             return null;
 
-        if (path.StartsWith(Application.dataPath)) {
-            return path.Substring(Application.dataPath.Length - "Assets".Length);
+        string slashedPath = path.Replace('\\', '/');        
+        string projectRoot = GetApplicationRootPath();
+        
+        if (slashedPath.StartsWith(projectRoot)) {
+            string normalizedPath = slashedPath.Substring(projectRoot.Length);
+            if (normalizedPath.Length > 0) {
+                normalizedPath = normalizedPath.Substring(1); //1 for additional '/'           
+            }
+
+            return normalizedPath;
         }
-        return path;
+        return slashedPath;
     }
+
+//----------------------------------------------------------------------------------------------------------------------
+
+    /// <summary>
+    /// Returns whether the path points to a path under "Assets" folder
+    /// </summary>
+    public static bool IsAssetPath(string path) {
+        string normalizedPath = NormalizeAssetPath(path);        
+        string[] dirs = normalizedPath.Split('/');
+        return (dirs.Length > 0 && dirs[0] == "Assets");
+    }
+    
+//----------------------------------------------------------------------------------------------------------------------    
+
+    static string GetApplicationRootPath() {
+        if (null != m_appRootPath)
+            return m_appRootPath;
+        
+        m_appRootPath = PathUtility.GetDirectoryName(Application.dataPath).Replace('\\','/');
+        return m_appRootPath;
+    }
+
+//----------------------------------------------------------------------------------------------------------------------    
+
+    private static string m_appRootPath = null;
 
 }